Use uint16_t for our various header sizes in bytes
PacketBuffer uses uint16_t for all of its sizes already.  It's easier
to tell that we're not overflowing the buffer if we know our header
sizes also fit in uint16_t.

+++ src/transport/tests/TestSecurePairingSession.cpp
@@ -44,7 +44,7 @@ public:
         if (peer != nullptr)
         {
             MessageHeader hdr;
-            size_t headerSize = 0;
+            uint16_t headerSize = 0;
 
             hdr.packetHeader.Decode(msgBuf->Start(), msgBuf->DataLength(), &headerSize);
             msgBuf->ConsumeHead(headerSize);
